+++ src/main/java/opendct/tuning/discovery/discoverers/HDHomeRunDiscoverer.java
@@ -87,7 +87,7 @@ public class HDHomeRunDiscoverer implements DeviceDiscoverer {
         while (true) {
             try {
                 streamingWait = new LongDeviceOption(
-                        Config.getInteger("hdhr.wait_for_streaming", 15000),
+                        Config.getLong("hdhr.wait_for_streaming", 15000),
                         false,
                         "Return to SageTV",
                         "hdhr.wait_for_streaming",
